pyproject.toml
[tool.poetry]
name = "hospitalgpt"
version = "0.1.0"
description = ""
authors = ["Mick Lynch <mick.lynch@gmail.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyautogen = "^0.2.3"
requests = "^2.31.0"
openai = "^1.10.0"
python-dotenv = "^1.0.1"
python-dateutil = "^2.8.2"
langgraph = "*"
openrouter = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
